Refactoring.

+++ app/src/main/java/com/example/currencyconverter/data/remote/CurrencyConversionDto.kt
@@ -6,6 +6,6 @@ data class CurrencyConversionDto(
     @SerializedName("from") val from: String? = null,
     @SerializedName("to") val to: String? = null,
     @SerializedName("rate") val rate: Double? = null,
-    @SerializedName("fromAmount") val fromAmount: Double? = null,
-    @SerializedName("toAmount") val toAmount: Double? = null
+    @SerializedName("fromAmount") val fromAmount: Float? = null,
+    @SerializedName("toAmount") val toAmount: Float? = null
 )
